// frontend/src/types/generated/BindingBoxTree.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BindingBoxTreeNode } from "./BindingBoxTreeNode";

export type BindingBoxTree = {
  nodes: Array<BindingBoxTreeNode>;
  sizeConstraints: Array<[[number, number], [number | null, number | null]]>;
};